Moves NetworkingConfig to androidTest package Temporary Fix for ApiServiceTest
