+++ app/src/androidTest/java/io/github/farhad/popcorn/remote/NetworkingConfig.kt
@@ -1,10 +1,10 @@
-package io.github.farhad.popcorn.data.remote.api
+package io.github.farhad.popcorn.remote
 
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
+import io.github.farhad.popcorn.data.remote.api.DateTypeConverter
 import okhttp3.Interceptor
 import okhttp3.OkHttpClient
-import okhttp3.Response
 import retrofit2.Retrofit
 import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
 import retrofit2.converter.gson.GsonConverterFactory
@@ -53,17 +53,3 @@ class MovieApiHttpClient {
     }
 }
 
-class ApiKeyInterceptor(private val apiKey: String) : Interceptor {
-    override fun intercept(chain: Interceptor.Chain): Response {
-        var request = chain.request()
-
-        val url = request.url()
-            .newBuilder()
-            .addQueryParameter("api_key", apiKey)
-            .build()
-
-        request = request.newBuilder().url(url).build()
-
-        return chain.proceed(request)
-    }
-}
